[scheduler] Bug Fix: Fix issue with direct mapping id collisions
